Bugfix: Major performance improvement by fixing AI_Wait always waiting 0 seconds
AI_Wait will be used by many script states, such as ZS_STAY. Since executing
scripts is expensive, it wants to wait for 1s before executing the next cycle.
AI_Wait however gets the `seconds`-argument as float. Floats are extremely
weird in Gothic as they are encoded as integers on the stack. Since the PushInt
instruction is used to push the seconds value, it will not end up on the
float-stack, and this AI_Wait reads 0.

I have profiled that to take about 40% of execution time in OLDMINE.ZEN
in a non-optimized build.